--- Лабораторная 0/Readme.md
@@ -1,4 +1,0 @@
-## Визитка 
-**ФИО** - Чуменко Игорь
-
-**Группа** - ПИНб-21
