--- Лабораторные/Лабораторная 2/interface.py
@@ -1,18 +1,0 @@
-class ArrayInterface:
-    def find(self, search_value):
-        ...
-
-    def insert(self, value):
-        ...
-    
-    def delete(self, value):
-        ...
-    
-    def display(self):
-        ...
-
-    def get_min(self):
-        ...
-
-    def get_max(self):
-        ...
