+++ Лабораторные/Первый семестр/Лабораторная 3/interface.py
@@ -1,0 +1,18 @@
+class ArrayInterface:
+    def find(self, search_value):
+        ...
+
+    def insert(self, value):
+        ...
+    
+    def delete(self, value):
+        ...
+    
+    def display(self):
+        ...
+
+    def get_min(self):
+        ...
+
+    def get_max(self):
+        ...
