+++ Лабораторные/Первый семестр/Лабораторная 5/mane.py
